backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
python-multipart
pydantic-settings
langchain
langchain-core
langchain-groq